fix(util): renderer function move to sub file
